refactor(mlx): split message parsing out of Session.swift
Session.swift grew past 800 lines with the Phase 3 VLM work. None of
the new helpers — parseContent, decodeImageURL, fetchImageBytes,
decodeDataURL — are session-specific; they're pure JSON-to-Swift
conversion plus a synchronous HTTP fetch. Move them and the existing
decodeMessages / decodeTools into MessageParsing.swift alongside
DecodedMessages. Session.swift is now 562 lines and purely about the
FFI boundary plus the generation driver.

Drops the CoreImage import from Session.swift and adds it to
MessageParsing.swift where CIImage is actually referenced. Also adds
two doc clarifications: tool-result images are dropped (because
Chat.Message.tool(_:) has no images parameter), and the
instructions-empty-to-nil coalescing is explained as a consequence of
ChatSession.instructions being String?.
